feat: マッパーをdiする
