+++ src/main/kotlin/com/example/project/application/mapper/user/UserMapper.kt
@@ -3,7 +3,9 @@ package com.example.project.application.mapper.user
 import com.example.project.application.dto.user.UserDTO
 import com.example.project.application.dto.user.UserFormDTO
 import com.example.project.domain.model.user.User
+import org.springframework.stereotype.Component
 
+@Component
 class UserMapper {
   fun toDomainEntity(userFormDTO: UserFormDTO): User {
     return User(
