+++ src/main/kotlin/com/example/project/domain/exception/ApplicationError.kt
@@ -1,0 +1,8 @@
+package com.example.project.domain.exception
+
+data class ApplicationError(
+  val message: String = "",
+  val httpStatus: Int,
+  val httpMethod: String,
+  val fieldErrors: List<FieldError?> = listOf<FieldError>()
+)
