+++ src/main/kotlin/com/example/project/domain/exception/FieldError.kt
@@ -1,0 +1,6 @@
+package com.example.project.domain.exception
+
+data class FieldError(
+  val fieldName: String,
+  val message: String,
+)
