--- src/main/kotlin/com/example/project/error/ApplicationError.kt
@@ -1,8 +1,0 @@
-package com.example.project.error
-
-data class ApplicationError(
-    val message: String = "",
-    val httpStatus: Int,
-    val httpMethod: String,
-    val fieldErrors: List<FieldError?> = listOf<FieldError>()
-)
