--- src/main/kotlin/com/example/project/error/FieldError.kt
@@ -1,6 +1,0 @@
-package com.example.project.error
-
-data class FieldError(
-    val fieldName: String,
-    val message: String,
-)
