+++ src/main/kotlin/com/example/project/presentation/exception/exception/InvalidTokenException.kt
@@ -1,3 +1,3 @@
-package com.example.project.exception
+package com.example.project.presentation.exception.exception
 
 class InvalidTokenException(val httpStatusCode: Int, message: String? = null) : RuntimeException(message)
