+++ src/main/kotlin/com/example/project/presentation/exception/exception/TitleExistException.kt
@@ -1,3 +1,3 @@
-package com.example.project.exception
+package com.example.project.presentation.exception.exception
 
 class TitleExistException(val httpStatusCode: Int, message: String? = null) : RuntimeException(message)
