+++ src/main/kotlin/com/example/project/presentation/exception/exception/UserNotExistsException.kt
@@ -1,3 +1,3 @@
-package com.example.project.exception
+package com.example.project.presentation.exception.exception
 
 class UserNotExistsException(val httpStatusCode: Int, message: String? = null) : RuntimeException(message)
