+++ src/main/kotlin/com/example/project/presentation/mapper/user/UserFormMapper.kt
@@ -2,8 +2,10 @@ package com.example.project.presentation.mapper.user
 
 import com.example.project.application.dto.user.UserFormDTO
 import com.example.project.presentation.form.user.UserForm
+import org.springframework.stereotype.Component
 
-class UserMapper {
+@Component
+class UserFormMapper {
   fun toDto(userForm: UserForm): UserFormDTO {
     return UserFormDTO(
       userForm.userId,
